feat: create store for editor initialization
